deployment/customer_purchase_prediction/requirements.txt
numpy
pandas
scikit-learn
sdv
tensorflow
tensorflow_io
python-dotenv
django
matplotlib
seaborn
fastparquet
sqlalchemy
pydash
django-rest-framework